Don't overwrite existing zshrc.local

+++ install.sh
@@ -66,7 +66,8 @@ ln -fs ${REPO}/dot.tmux.conf ~/.tmux.conf
 ln -fs ${REPO}/dot.valgrindrc ~/.valgrindrc
 ln -fs ${REPO}/dot.vimrc ~/.vimrc
 ln -fs ${REPO}/dot.zlogout ~/.zlogout
-if [ -f ~/.zshrc ] && [ ! -L ~/.zshrc ]; then
+if [ -f ~/.zshrc ] && [ ! -L ~/.zshrc ] && [ ! -L ~/.zshrc.local ] &&
+    [ ! -f ~/.zshrc.local ]; then
 	mv ~/.zshrc ~/.zshrc.local
 fi
 ln -fs ${REPO}/dot.zshrc ~/.zshrc
